[WIP] Separating strings with distinct typedefs
By definition, `typedef` is strictly an alias for a type. Therefore,
the following sequence mean that `IdentifierName` and `Identifier`
are interchangeable:
```
typedef string IdentifierName;
typedef string Identifier;
```

As this is not true, the current patch is a first attempt to remove this aliasing.

We solve this by introducing an annotation `[Token(...)]`, which lets us specify
the token-space in which an alias of string belongs.

So,

```
typedef [Token(IdentifierName)] string IdentifierName;
typedef [Token(Identifier)] string Identifier;
```

specifies that an `IdentifierName` is a string from the set of tokens `IdentifierName`.

This is an early draft for feedback, we haven't yet actually defined the set of tokens.